refactor view api helper functions to separate service file
